encore des bugs

+++ lib/ft_lstclear.c
@@ -6,13 +6,13 @@
 /*   By: Laubry <aubrylucas.pro@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/26 10:51:38 by Laubry            #+#    #+#             */
-/*   Updated: 2023/12/11 16:00:20 by Laubry           ###   ########.fr       */
+/*   Updated: 2024/01/02 14:36:27 by Laubry           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_lstclear(t_list **lst, void (*del)(int))
+void	ft_lstclear(t_list **lst, void (*del)(long))
 {
 	t_list	*dest;
