+++ lib/ft_lstdelone.c
@@ -6,13 +6,13 @@
 /*   By: Laubry <aubrylucas.pro@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/26 10:35:28 by Laubry            #+#    #+#             */
-/*   Updated: 2023/12/11 16:02:24 by Laubry           ###   ########.fr       */
+/*   Updated: 2024/01/02 14:36:32 by Laubry           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_lstdelone(t_list *lst, void (*del)(int))
+void	ft_lstdelone(t_list *lst, void (*del)(long))
 {
 	if (!lst ||!del)
 		return ;
